+++ frontend/src/features/map_planning/layers/plant/components/PlantLayerRelationsOverlay.tsx
@@ -2,7 +2,7 @@ import { useRelations } from '../hooks/useRelations';
 import { LayerType, RelationType } from '@/api_types/definitions';
 import useMapStore from '@/features/map_planning/store/MapStore';
 import { useEffect, useMemo, useState } from 'react';
-import { Layer, Line } from 'react-konva';
+import { Group, Line } from 'react-konva';
 
 const relationColors = {
   [RelationType.Antagonist]: '#f005',
@@ -66,7 +66,7 @@ export function PlantLayerRelationsOverlay() {
   }, [data, layers]);
 
   return (
-    <Layer listening={false}>
+    <Group listening={false}>
       {!isLoading && lineEnd && relatedVisiblePlantings
         ? relatedVisiblePlantings.map((s) => {
             const relation = data?.get(s.attrs.plantId)?.relation;
@@ -82,6 +82,6 @@ export function PlantLayerRelationsOverlay() {
             );
           })
         : null}
-    </Layer>
+    </Group>
   );
 }